ai_print_estimator/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
python-multipart
python-dotenv
openai
tenacity
httpx
pymupdf
pillow
requests